refactor: (#230) 메소드 변경
